Fixed Equipment duplications
Check a diff between this and the previous commit to see what what added (ask me for git questions if you have them), this commit just adds a Singleton equipmetnQueue that keeps a sorted list of the equipment to be added. I changed how a few methods work in order to *attempt* to keep these changes safe.